requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
openai>=1.55.0
python-dotenv>=1.0.1
requests>=2.32.2
PyYAML>=6.0.1
httpx>=0.27.0
